拆分画布节点组件通过体积门岗
