+++ src/workbench/generationCanvas/nodes/whiteboard/WhiteboardModal.tsx
@@ -8,7 +8,8 @@ import { FULLSCREEN_Z_INDEX } from '../scene3d/scene3dConstants'
 import type { GenerationCanvasEdge, GenerationCanvasNode, GenerationNodeResult } from '../../model/generationCanvasTypes'
 import { useGenerationCanvasStore } from '../../store/generationCanvasStore'
 import { persistNodeImageFile } from '../../adapters/persistNodeImage'
-import WhiteboardDrawingTool, { type WhiteboardDrawingToolHandle, type WhiteboardResultLibraryItem } from './WhiteboardDrawingTool'
+import WhiteboardDrawingTool, { type WhiteboardDrawingToolHandle } from './WhiteboardDrawingTool'
+import type { WhiteboardResultLibraryItem } from './whiteboardTypes'
 import type { WhiteboardInitialImage, WhiteboardState } from './whiteboardTypes'
 import { readWhiteboardState, serializeWhiteboardState } from './whiteboardState'
 import { getCanvasDimensions } from './lib/canvas'
